Modify main source file to not include functionalities moved to namespaces. Part of https://stgit.dcs.gla.ac.uk/team-project-h/2021/cs34/cs34-main/-/issues/85
The functionality remains the same, but it makes the structure and further development a lot easier.
Main setup now expects the user to implement experiment namespace containing a begin and measure functions.
The function begin does not return anything and initialises the sensor itself.
The function measure is called periodically in the main loop and should take a single measurement and print a line to the serial with the following structure:
	data: value <units>